Adding Appstream metadata for applicaiton stores
This adds Appstream metadata for display in application stores on linux.
This closes https://github.com/EUA/wxHexEditor/pull/173 , and in
contrast to that is compatible with the existing appstream metadata
already on flathub.

The screenshots in the metadata will only begin to work when this is
merged into the main github upstream, because they are referenced to
there.

As it is here, changing the screenshots on master will automatically
change them in application stores.